backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
python-dotenv==1.0.1
gunicorn==21.2.0
pydantic==2.6.1
email-validator==2.1.0.post1
python-multipart==0.0.9
starlette==0.37.2

